Create a dedicated webhook for DELETE operations
Most constraints templates are design for CREATE and UPDATE operations,
but no for DELETE. Also, running all constraints on DELETE operations
means that it's not possible to delete pods that are violating a new
constraint.

Our interest in the DELETE operation is for protecting us against
accidental deletion of namespaces, so it makes sense to only check the
constraints when deleting namespaces.