Fix gatekeeper manifests generation
Helm v3 includes a "capabilities" api to allow templating based on the
available apis. Sadly this functionality does not work with `helm
template` and all the checks fail, unless apiversions are explicitely
passed as flags.

To make the issue worse, helms checks the full api name (as in
policy/v1/PodDisruptionBudget) instead of just the group (as in
policy/v1), which is what kubernetes provides via
`kubectl api-versions`, so there's no easy way to programatically get
a the list.

As a result, I'm adding the manual API_VERSIONS variable, which we need
to maintain :(

On a side note, API_VERSIONS needs a trailing comma to be interpreted as
an array, which is what helm expects, now that it only has 1 element.