Localize Traefik CRDs
Traefik helm repo is now also a Kustomize base. Instead of using it
directly (in case it ever goes away, and the versioning does not
correspont Traefik versions anyway), we localize a version we know is
good for current "latest" Traefik.

We get predictable manifests that we retain control over and automate
update process in a Kuztomize native way.