Fix Kyverno webhooks and update deprecated fields
- Remove autoUpdateWebhooks=false flag to allow Kyverno create and edit its
  webhooks and set different failurePolicy for resources according to policies
- Update `validationFailureAction` values to start with capital letters

tldr; Setting autoUpdateWebhooks=false will trigger Kyverno to use
`failurePolicy: Ignore` in all the webhooks it creates which means that even if
an admission request is denied by Kyverno the apiserver will still accept it.
Changing that will make the webhook setup behave as expected, but will result in
admissions being rejected in cases where Kyverno is down and cannot reply to
apiserver.