cloud-provider-gcp: Initial deploy of kustomize base
This is creating a target to fetch from an upstream example kops deployment and
patches arguments, volumes and image to give us a base to deploy
cloud-provider-gcp. More in detail:
- For some reason upstream manifests require hostNetwork, which is not a kops
things but seems to be involved in all their manifest templates.
- Disables controllers we do not need due to calico, based on this:
https://github.com/kubernetes/cloud-provider-gcp/pull/482#issuecomment-1450197489
to make cloud-provider not deal with setting up routes for pod networking, or
GKE related stuff
- Patches system::leader-locking-cloud-controller-manager role for migration
- Allows controller-manager to manage nodes