Add manifest for OpsLevel alias injector
See README for details on what this policy does. For the moment this
policy will only act under the `dev-enablement` namespace just to see
how things go.

This is being added here, rather than in our mono-repo to keep all the
kyverno policies in one place, so we don't have to clone another repo to
build these manifests

I've already deployed a version of this in `dev-aws` (see [1],[2] I'll
remove this shortly) which after updating a deployment by just adding
then removing a new annotation, creates pods having (note: I've since
changed the name of the annotation/env var so it's different to what's
in this diff):

    $ kubectl --context dev-aws --namespace dev-enablement get pod --output json opslevel-k8s-deployer-6fb76f676b-86zss | jq '.metadata.annotations."app.uw.systems/opslevel-service"'
    "k8s:deployment:dev-enablement-opslevel-k8s-deployer"
    $ kubectl --context dev-aws --namespace dev-enablement get pod --output json opslevel-k8s-deployer-6fb76f676b-86zss | jq '.spec.containers[].env[] | select(.name=="OPSLEVEL_SERVICE")'
    {
      "name": "OPSLEVEL_SERVICE",
      "value": "k8s:deployment:dev-enablement-opslevel-k8s-deployer"
    }

[1] https://github.com/utilitywarehouse/kubernetes-manifests/pull/74533
[2] https://github.com/utilitywarehouse/kubernetes-manifests/pull/74538

Ticket: DENA-149