Revert "Add manifest for OpsLevel alias injector" (#219)
This reverts commit 771d015e13076015f7b7ed7c86fdfb1648bbbaad. Two
reasons for this: since we decided to not proceed with this approach
across all namespaces (see 3a3a845fd02a13b6d29334f49251e930d8472b8c)
there's not much need for it to run in dev-enablement.

Secondly there was a bug caused by an earlier iteration of this
injector[1]: one of our systems appears to try and load config from
environment variables prefixed with `OPSLEVEL`, specifically we have an
application[2] that looks to load config for `service` which we provide
at[3], however since this application runs in `dev-aws` with this
injector, the env var `OPSLEVEL_SERVICE` was set, which the program than
tried to read from instead and hence loaded an empty config. The config
is used to define which k8s resources to reconcile, since the empty
config matched nothing the application would just sit waiting and filter
out all events and never actually act. For this bug I've just manually
edited the deployment to drop the env var that was injected by the old
implementation, but I think it's worthwhile to just drop this policy
entirely to avoid any potential conflicts/ambiguity in the future.

[1] https://github.com/utilitywarehouse/kubernetes-manifests/pull/74533
[2] https://github.com/OpsLevel/kubectl-opslevel
[3] https://github.com/utilitywarehouse/dev-enablement-mono/blob/a51f49aaa4ba8974a34155728a57809ae117eb0a/services/opslevel-k8s-reconciler/k8s-manifests/base/config/config.yaml#L2

Ticket: DENA-220